refactor: 카드 계산 로직 분리하기
